Rewrote tests using Jest
